Moved eye constructor out to dedicated implementation file
